Add: Ability to init the dataabse at startup.
Add the ability to init the db at startup.
